Fix imports
This commit moves out the GenericTable definition to a separate file and
rewrites SQLAlchemy and FAF imports to import from the actual modules.
The former makes pylint warnings about cyclic imports go away and the
latter helps with readability in general.
