+++ src/pyfaf/storage/bugtracker.py
@@ -16,12 +16,12 @@
 # You should have received a copy of the GNU General Public License
 # along with faf.  If not, see <http://www.gnu.org/licenses/>.
 
+from sqlalchemy.sql.schema import Column
+from sqlalchemy.types import Integer, String
+
 from pyfaf.config import config
 
-from . import Column
-from . import GenericTable
-from . import Integer
-from . import String
+from .generic_table import GenericTable
 
 
 class Bugtracker(GenericTable):
