+++ src/pyfaf/storage/project.py
@@ -16,13 +16,11 @@
 # You should have received a copy of the GNU General Public License
 # along with faf.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import Column
-from . import DateTime
-from . import ForeignKey
-from . import GenericTable
-from . import Integer
-from . import String
-from . import relationship
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql.schema import Column, ForeignKey
+from sqlalchemy.types import DateTime, Integer, String
+
+from .generic_table import GenericTable
 
 
 class Project(GenericTable):
