+++ src/pyfaf/storage/sf_prefilter.py
@@ -16,14 +16,12 @@
 # You should have received a copy of the GNU General Public License
 # along with faf.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import Column
-from . import ForeignKey
-from . import GenericTable
-from . import Integer
-from . import OpSys
-from . import String
-from . import UniqueConstraint
-from . import relationship
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql.schema import Column, ForeignKey, UniqueConstraint
+from sqlalchemy.types import Integer, String
+
+from .generic_table import GenericTable
+from .opsys import OpSys
 
 
 class SfPrefilterSolution(GenericTable):
