+++ src/pyfaf/storage/symbol.py
@@ -16,14 +16,11 @@
 # You should have received a copy of the GNU General Public License
 # along with faf.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import BigInteger
-from . import Column
-from . import ForeignKey
-from . import GenericTable
-from . import Integer
-from . import String
-from . import UniqueConstraint
-from . import relationship
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql.schema import Column, ForeignKey, UniqueConstraint
+from sqlalchemy.types import BigInteger, Integer, String
+
+from .generic_table import GenericTable
 
 
 class Symbol(GenericTable):
